agents/scraping_agent/requirements-scraping.txt
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
langchain>=0.0.335
unstructured>=0.10.30
beautifulsoup4>=4.12.2
lxml>=4.9.3
